+++ tests/api/mod.rs
@@ -1,1 +1,7 @@
-pub mod two;
+pub mod blocks_test;
+pub mod delegates_test;
+pub mod node_test;
+pub mod peers_test;
+pub mod transactions_test;
+pub mod votes_test;
+pub mod wallets_tests;
