+++ tests/api/peers_test.rs
